+++ spring-cloud-loadbalancer/src/main/java/org/springframework/cloud/loadbalancer/core/LoadBalancer.java
@@ -14,11 +14,36 @@
  * limitations under the License.
  */
 
-package org.springframework.cloud.client.loadbalancer.impl;
+package org.springframework.cloud.loadbalancer.core;
 
 /**
  * @author Spencer Gibb
  */
-class DefaultRequest implements LoadBalancer.Request {
+public interface LoadBalancer<T> {
 
+	/**
+	 * Response created for each request.
+ 	 */
+	interface Response<T> {
+		boolean hasServer();
+
+		T getServer();
+
+		/**
+		 * Notification that the request completed
+		 * @param onComplete
+		 */
+		void onComplete(OnComplete onComplete);
+	}
+
+	interface Request {
+		//TODO: define contents
+	}
+
+	/**
+	 * Choose the next server based on the load balancing algorithm
+	 * @param request
+	 * @return
+	 */
+	Response<T> choose(Request request);
 }
